+++ aesqlapius/function_def.py
@@ -45,6 +45,7 @@ class ReturnValueInnerFormat(Enum):
     TUPLE = 1
     DICT = 2
     LIST = 3
+    SINGLE = 4
 
 
 @dataclass
@@ -72,6 +73,8 @@ def _parse_return_value_inner(node: ast.AST) -> Union[ReturnValueInnerFormat, st
         return ReturnValueInnerFormat.DICT
     elif node.id == 'List':
         return ReturnValueInnerFormat.LIST
+    elif node.id == 'Single':
+        return ReturnValueInnerFormat.SINGLE
     else:  # custom type
         return node.id