+++ tests/test_postgresql/get/__init__.sql
@@ -10,6 +10,9 @@ SELECT 0 AS a, 'a' AS b;
 -- def single_tuple() -> Single[Tuple]: ...
 SELECT 0 AS a, 'a' AS b;
 
+-- def single_single() -> Single[Single]: ...
+SELECT 0 AS a;
+
 -- def iterator_dict() -> Iterator[Dict]: ...
 SELECT a, chr(ascii('a') + a) AS b FROM generate_series(0, 2) a;
 
@@ -19,6 +22,9 @@ SELECT a, chr(ascii('a') + a) AS b FROM generate_series(0, 2) a;
 -- def iterator_tuple() -> Iterator[Tuple]: ...
 SELECT a, chr(ascii('a') + a) AS b FROM generate_series(0, 2) a;
 
+-- def iterator_single() -> Iterator[Single]: ...
+SELECT a FROM generate_series(0, 2) a;
+
 -- def list_dict() -> List[Dict]: ...
 SELECT a, chr(ascii('a') + a) AS b FROM generate_series(0, 2) a;
 
@@ -26,6 +32,9 @@ SELECT a, chr(ascii('a') + a) AS b FROM generate_series(0, 2) a;
 SELECT a, chr(ascii('a') + a) AS b FROM generate_series(0, 2) a;
 
 -- def list_tuple() -> List[Tuple]: ...
+SELECT a, chr(ascii('a') + a) AS b FROM generate_series(0, 2) a;
+
+-- def list_single() -> List[Single]: ...
 SELECT a, chr(ascii('a') + a) AS b FROM generate_series(0, 2) a;
 
 -- def dict_of_tuples_by_column_name() -> Dict['a', Tuple]: ...